perf/JobPool: make allocate() return a job then use placement new in the function[] field
Originally jobs were allocated then the function object was memcpy'd into the `function` field of the job. By allocating the job and then calling placement new in the already existing memory in the `function` field, we avoid an allocation and copy of the function data and reduce the run time of allocate() by up to 2/3.